docs: Improve upgrading docs
* The edit commands had the namespace, but it was missing in the set-image
variant. This patch just adds the namespace for the set-image command.

* kube-proxy can also be updated with set-image, so let's add the example.

* When running `kubectl get nodes` the old version will be shown until the
kubelets are updated. Therefore, this patch moves that after the kubelet
was updated and suggest verifying that upgraded pods are running
instead.

 * The code is already using k8s.gcr.io/hyperkube. This patch just updates
the documentation to use that too. I've checked and the images used for
the example (k8s 1.6.7 and 1.6.6) exist under the upstream repo, so kept
that version in the example while just removing the coreos part (so, the
git commit in the examples is not correct, didn't think it is a big
issue worth fixing).