fix: remove node selector from static manifests
Due to changes in Kubernetes 1.16, the node-role.kubernetes.io label can
no longer be automatically applied by the kubelet. Due to this, we had
to remove it from the `--node-labels` flag. This caused the static
pod-checkpointer to fail to run after a complete loss of the control
plane (i.e. all control plane nodes rebooted) because it had a
`nodeSelector` inherited from the `DaemonSet` that used the
`node-role.kubernetes.io` label. This PR adds logic to strip off all
`node-role.kubernetes.io` related metadata from the static manifest, which
allows the scheduling to occur.

Signed-off-by: Andrew Rynhard <andrew@andrewrynhard.com>